Выносит offers-container в отдельный компонент
