+++ src/view/offers-container.js
@@ -1,0 +1,19 @@
+export const renderOffersContainer = (offers) => {
+  const renderOffers = (offers) => {
+    return  offers ? offers.map((offer) => `<div class="event__offer-selector">
+      <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.title}-1" type="checkbox" name="event-offer-${offer.title}" checked>
+      <label class="event__offer-label" for="event-offer-${offer.title}-1">
+        <span class="event__offer-title">${offer.title}</span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price">${offer.price}</span>
+      </label>
+    </div>`)
+      .join('') : offers = ' ';
+  };
+  return offers ?  `<section class="event__section  event__section--offers">
+    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+    <div class="event__available-offers">
+    ${renderOffers(offers)}
+    </div>
+    </section>` : '';
+};
